requirements.txt
# requirements.txt
streamlit==1.39.0
langchain==0.3.7
langchain-google-vertexai==2.0.8
google-cloud-aiplatform==1.70.0
google-cloud-storage==2.18.2
pydantic==2.9.2
plotly==5.24.1
python-dotenv==1.0.1
langchain-community==0.3.5
tavily-python==0.3.3